# backend/docs/swagger.yaml
definitions:
...
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      password:
        description: パスワード
        example: password
        type: string
      username:
        description: ユーザー名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handler.LoginResponse:
    properties:
      message:
        description: メッセージ
        example: ログインに成功しました
        type: string
      role:
        description: ユーザーロール
        example: admin
        type: string
      success:
        description: 成功フラグ
        example: true
        type: boolean
      token:
        description: JWTトークン
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      username:
        description: ユーザー名
        example: admin
        type: string
    type: object
...
  handler.RefreshTokenRequest:
    properties:
      token:
        description: 既存のJWTトークン
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - token
    type: object
  handler.RefreshTokenResponse:
    properties:
      message:
        description: メッセージ
        example: トークンのリフレッシュに成功しました
        type: string
      success:
        description: 成功フラグ
        example: true
        type: boolean
      token:
        description: 新しいJWTトークン
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 管理者認証情報でログインし、JWTトークンを取得する
      parameters:
      - description: ログイン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ログイン成功
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: サーバーエラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: ユーザーログイン
      tags:
      - auth
...
  /api/auth/profile:
    get:
      description: 現在認証されているユーザーの情報を取得する
      produces:
      - application/json
      responses:
        "200":
          description: ユーザー情報
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ユーザープロフィール取得
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: 既存のJWTトークンを使用して新しいトークンを生成する
      parameters:
      - description: リフレッシュ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: リフレッシュ成功
          schema:
            $ref: '#/definitions/handler.RefreshTokenResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: サーバーエラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: JWTトークンリフレッシュ
      tags:
      - auth
  /api/auth/validate:
    post:
      consumes:
      - application/json
      description: JWTトークンの有効性を検証する
      parameters:
      - description: 検証するトークン
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 検証成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: トークン検証
      tags:
      - auth
...
swagger: "2.0"